Replace the moderation gatedComponent HOC with a single runtime gate
The enterprise moderation plugin gated each member individually — a
gatedComponent HOC wrapping every component plus per-function isEnabled()
checks. Collapse that into one idea: when content_verification is absent, the
plugin behaves as the OSS default.

- Add a reusable gatedPlugin(isEnabled, enabled, disabled) helper that returns a
  Proxy delegating every member to the enabled or disabled implementation,
  evaluated lazily (token features load after the bundle evaluates).
- Split the OSS default into metabase/moderation/default.ts so the enterprise
  build can import it as the disabled fallback without the
  metabase/moderation/plugin build-override pointing the import back at itself;
  plugin.ts becomes the resolved entry that re-exports it.
- The enterprise implementation is now a plain object (no HOC, no per-member
  gating); the file no longer needs JSX, so it is plugin.ts not plugin.tsx.
